Data Insertion:
Insert a substantial amount of data into the public data form. Modify the React components to efficiently utilize and accurately represent the data. The current data size is causing the program to slow down and lag. Note that NoSQL implementation is imminent, and all data will be stored there in the near future.